changed filenames to indicate classes
