+++ src/controllers/AnimeController.js
@@ -1,30 +1,30 @@
 // Import the neccesary modules.
-import Movie from "../models/Movie";
+import Anime from "../models/Anime";
 import { pageSize } from "../config/constants";
 
-/** Class for getting movie data from the MongoDB. */
-export default class Movies {
+/** Class for getting anime data from the MongoDB. */
+export default class AnimeController {
 
-  /** Create a movies object. */
+  /** Create an anime controller object. */
   constructor() {
     /**
-     * Object used for the projection of movies.
+     * Object used for the projection of anime shows.
      * @type {Object}
      */
-    Movies._projection = {
-      _id: 1,
+    AnimeController._projection = {
+      images: 1,
+      mal_id: 1,
+      haru_id: 1,
+      tvdb_id: 1,
       imdb_id: 1,
+      slug: 1,
       title: 1,
       year: 1,
-      runtime: 1,
-      images: 1,
-      genres: 1,
-      synopsis: 1,
-      trailer: 1,
-      certification: 1,
-      released: 1,
+      type: 1,
+      item_data: 1,
       rating: 1,
-      torrents: 1
+      genres: 1,
+      num_seasons: 1
     };
   }
 
@@ -35,12 +35,17 @@ export default class Movies {
    * @param {Function} next - The next function for Express.
    * @returns {String[]} - A list of pages which are available.
    */
-  getMovies(req, res, next) {
-    return Movie.count().exec().then(count => {
+  getAnimes(req, res, next) {
+    return Anime.count({
+      num_seasons: {
+        $gt: 0
+      },
+      type: "show"
+    }).exec().then(count => {
       const pages = Math.round(count / pageSize);
       const docs = [];
 
-      for (let i = 1; i < pages + 1; i++) docs.push(`movies/${i}`);
+      for (let i = 1; i < pages + 1; i++) docs.push(`animes/${i}`);
 
       return res.json(docs);
     }).catch(err => next(err));
@@ -51,15 +56,22 @@ export default class Movies {
    * @param {Request} req - The express request object.
    * @param {Response} res - The express response object.
    * @param {Function} next - The next function for Express.
-   * @returns {Movie[]} - The contents of one page.
+   * @returns {Anime[]} - The contents of one page.
    */
   getPage(req, res, next) {
     const page = req.params.page - 1;
     const offset = page * pageSize;
 
     if (req.params.page.match(/all/i)) {
-      return Movie.aggregate([{
-          $project: Movies._projection
+      return Anime.aggregate([{
+          $match: {
+            num_seasons: {
+              $gt: 0
+            },
+            type: "show"
+          }
+        }, {
+          $project: AnimeController._projection
         }, {
           $sort: {
             title: -1
@@ -68,7 +80,12 @@ export default class Movies {
         .then(docs => res.json(docs))
         .catch(err => next(err));
     } else {
-      const query = {};
+      const query = {
+        num_seasons: {
+          $gt: 0
+        },
+        type: "show"
+      };
       const data = req.query;
 
       if (!data.order) data.order = -1;
@@ -95,35 +112,26 @@ export default class Movies {
       }
 
       if (data.sort) {
-        if (data.sort.match(/last added/i)) sort = {
-          "released": parseInt(data.order, 10)
+        if (data.sort.match(/name/i)) sort = {
+          "title": (parseInt(data.order, 10) * -1)
         };
         if (data.sort.match(/rating/i)) sort = {
           "rating.percentage": parseInt(data.order, 10),
           "rating.votes": parseInt(data.order, 10)
-        };
-        if (data.sort.match(/title/i)) sort = {
-          "title": (parseInt(data.order, 10) * 1)
-        };
-        if (data.sort.match(/trending/i)) sort = {
-          "rating.watching": parseInt(data.order, 10)
         };
         if (data.sort.match(/year/i)) sort = {
           "year": parseInt(data.order, 10)
         };
       }
 
-      if (data.genre && !data.genre.match(/all/i)) {
-        if (data.genre.match(/science[-\s]fiction/i) || data.genre.match(/sci[-\s]fi/i)) data.genre = "science-fiction";
-        query.genres = data.genre.toLowerCase();
-      }
+      if (data.genre && !data.genre.match(/all/i)) query.genres = data.genre;
 
-      return Movie.aggregate([{
+      return Anime.aggregate([{
           $sort: sort
         }, {
           $match: query
         }, {
-          $project: Movies._projection
+          $project: AnimeController._projection
         }, {
           $skip: offset
         }, {
@@ -135,36 +143,37 @@ export default class Movies {
   }
 
   /**
-   * Get info from one movie.
+   * Get info from one anime.
    * @param {Request} req - The express request object.
    * @param {Response} res - The express response object.
    * @param {Function} next - The next function for Express.
-   * @returns {Movie} - The details of a single movie.
+   * @returns {Anime} - The details of a single anime.
    */
-  getMovie(req, res, next) {
-    return Movie.aggregate([{
-        $match: {
-          _id: req.params.id
-        }
+  getAnime(req, res, next) {
+    return Anime.findOne({
+        _id: req.params.id,
+        type: "show"
       }, {
-        $project: Movies._projection
-      }, {
-        $limit: 1
-      }]).exec()
-      .then(docs => res.json(docs[0]))
+        latest_episode: 0
+      }).exec()
+      .then(docs => res.json(docs))
       .catch(err => next(err));
   }
 
   /**
-   * Get a random movie.
+   * Get a random anime.
    * @param {Request} req - The express request object.
    * @param {Response} res - The express response object.
    * @param {Function} next - The next function for Express.
-   * @returns {Movie} - A random movie.
+   * @returns {Anime} - A random movie.
    */
-  getRandomMovie(req, res, next) {
-    return Movie.aggregate([{
-        $project: Movies._projection
+  getRandomAnime(req, res, next) {
+    return Anime.aggregate([{
+        $match: {
+          type: "show"
+        }
+      }, {
+        $project: AnimeController._projection
       }, {
         $sample: {
           size: 1
