+++ src/controllers/ExportController.js
@@ -2,13 +2,13 @@
 import fs from "fs";
 import path from "path";
 
-import Util from "../util";
+import Util from "../Util";
 import { tempDir } from "../config/constants";
 
 /** Class for getting anime data from the MongoDB. */
-export default class Exports {
+export default class ExportController {
 
-  /** Create an exports object. */
+  /** Create an export controller object. */
   constructor() {
     /**
      * The util object with general functions.
