+++ src/controllers/MovieController.js
@@ -1,30 +1,30 @@
 // Import the neccesary modules.
-import Anime from "../models/Anime";
+import Movie from "../models/Movie";
 import { pageSize } from "../config/constants";
 
-/** Class for getting anime data from the MongoDB. */
-export default class Animes {
+/** Class for getting movie data from the MongoDB. */
+export default class MovieController {
 
-  /** Create an animes object. */
+  /** Create a movie controller object. */
   constructor() {
     /**
-     * Object used for the projection of anime shows.
+     * Object used for the projection of movies.
      * @type {Object}
      */
-    Animes._projection = {
-      images: 1,
-      mal_id: 1,
-      haru_id: 1,
-      tvdb_id: 1,
+    MovieController._projection = {
+      _id: 1,
       imdb_id: 1,
-      slug: 1,
       title: 1,
       year: 1,
-      type: 1,
-      item_data: 1,
-      rating: 1,
+      runtime: 1,
+      images: 1,
       genres: 1,
-      num_seasons: 1
+      synopsis: 1,
+      trailer: 1,
+      certification: 1,
+      released: 1,
+      rating: 1,
+      torrents: 1
     };
   }
 
@@ -35,17 +35,12 @@ export default class Animes {
    * @param {Function} next - The next function for Express.
    * @returns {String[]} - A list of pages which are available.
    */
-  getAnimes(req, res, next) {
-    return Anime.count({
-      num_seasons: {
-        $gt: 0
-      },
-      type: "show"
-    }).exec().then(count => {
+  getMovies(req, res, next) {
+    return Movie.count().exec().then(count => {
       const pages = Math.round(count / pageSize);
       const docs = [];
 
-      for (let i = 1; i < pages + 1; i++) docs.push(`animes/${i}`);
+      for (let i = 1; i < pages + 1; i++) docs.push(`movies/${i}`);
 
       return res.json(docs);
     }).catch(err => next(err));
@@ -56,22 +51,15 @@ export default class Animes {
    * @param {Request} req - The express request object.
    * @param {Response} res - The express response object.
    * @param {Function} next - The next function for Express.
-   * @returns {Anime[]} - The contents of one page.
+   * @returns {Movie[]} - The contents of one page.
    */
   getPage(req, res, next) {
     const page = req.params.page - 1;
     const offset = page * pageSize;
 
     if (req.params.page.match(/all/i)) {
-      return Anime.aggregate([{
-          $match: {
-            num_seasons: {
-              $gt: 0
-            },
-            type: "show"
-          }
-        }, {
-          $project: Animes._projection
+      return Movie.aggregate([{
+          $project: MovieController._projection
         }, {
           $sort: {
             title: -1
@@ -80,12 +68,7 @@ export default class Animes {
         .then(docs => res.json(docs))
         .catch(err => next(err));
     } else {
-      const query = {
-        num_seasons: {
-          $gt: 0
-        },
-        type: "show"
-      };
+      const query = {};
       const data = req.query;
 
       if (!data.order) data.order = -1;
@@ -112,26 +95,35 @@ export default class Animes {
       }
 
       if (data.sort) {
-        if (data.sort.match(/name/i)) sort = {
-          "title": (parseInt(data.order, 10) * -1)
+        if (data.sort.match(/last added/i)) sort = {
+          "released": parseInt(data.order, 10)
         };
         if (data.sort.match(/rating/i)) sort = {
           "rating.percentage": parseInt(data.order, 10),
           "rating.votes": parseInt(data.order, 10)
+        };
+        if (data.sort.match(/title/i)) sort = {
+          "title": (parseInt(data.order, 10) * 1)
+        };
+        if (data.sort.match(/trending/i)) sort = {
+          "rating.watching": parseInt(data.order, 10)
         };
         if (data.sort.match(/year/i)) sort = {
           "year": parseInt(data.order, 10)
         };
       }
 
-      if (data.genre && !data.genre.match(/all/i)) query.genres = data.genre;
+      if (data.genre && !data.genre.match(/all/i)) {
+        if (data.genre.match(/science[-\s]fiction/i) || data.genre.match(/sci[-\s]fi/i)) data.genre = "science-fiction";
+        query.genres = data.genre.toLowerCase();
+      }
 
-      return Anime.aggregate([{
+      return Movie.aggregate([{
           $sort: sort
         }, {
           $match: query
         }, {
-          $project: Animes._projection
+          $project: MovieController._projection
         }, {
           $skip: offset
         }, {
@@ -143,37 +135,36 @@ export default class Animes {
   }
 
   /**
-   * Get info from one anime.
+   * Get info from one movie.
    * @param {Request} req - The express request object.
    * @param {Response} res - The express response object.
    * @param {Function} next - The next function for Express.
-   * @returns {Anime} - The details of a single anime.
+   * @returns {Movie} - The details of a single movie.
    */
-  getAnime(req, res, next) {
-    return Anime.findOne({
-        _id: req.params.id,
-        type: "show"
+  getMovie(req, res, next) {
+    return Movie.aggregate([{
+        $match: {
+          _id: req.params.id
+        }
       }, {
-        latest_episode: 0
-      }).exec()
-      .then(docs => res.json(docs))
+        $project: MovieController._projection
+      }, {
+        $limit: 1
+      }]).exec()
+      .then(docs => res.json(docs[0]))
       .catch(err => next(err));
   }
 
   /**
-   * Get a random anime.
+   * Get a random movie.
    * @param {Request} req - The express request object.
    * @param {Response} res - The express response object.
    * @param {Function} next - The next function for Express.
-   * @returns {Anime} - A random movie.
+   * @returns {Movie} - A random movie.
    */
-  getRandomAnime(req, res, next) {
-    return Anime.aggregate([{
-        $match: {
-          type: "show"
-        }
-      }, {
-        $project: Animes._projection
+  getRandomMovie(req, res, next) {
+    return Movie.aggregate([{
+        $project: MovieController._projection
       }, {
         $sample: {
           size: 1
