+++ src/popcorn-api.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 // Import the neccesary modules.
-import CLI from "./cli";
+import CLI from "./CLI";
 
 // Run the CLI program.
 new CLI().run();
