+++ src/providers/anime/HorribleSubs.js
@@ -2,8 +2,8 @@
 import asyncq from "async-q";
 import HorribleSubsAPI from "horriblesubs-api";
 
-import Extractor from "../extractors/animeextractor";
-import Util from "../../util";
+import Extractor from "../extractors/AnimeExtractor";
+import Util from "../../Util";
 import { maxWebRequest } from "../../config/constants";
 
 /** Class for scraping anime from https://horriblesubs.info/. */
