+++ src/providers/anime/KAT.js
@@ -2,15 +2,15 @@
 import asyncq from "async-q";
 import KatAPI from "kat-api-pt";
 
-import Extractor from "../extractors/showextractor";
-import Util from "../../util";
+import Extractor from "../extractors/AnimeExtractor";
+import Util from "../../Util";
 
-/** Class for scraping shows from https://kat.cr/. */
+/** Class for scraping anime shows from https://kat.cr/. */
 export default class KAT {
 
   /**
-   * Create a kat object for show content.
-   * @param {String} name - The name of the torrent provider.
+   * Create a kat object for anime content.
+   * @param {String} name - The name of the content provider.
    * @param {?Boolean} debug - Debug mode for extra output.
    */
   constructor(name, debug) {
@@ -21,7 +21,7 @@ export default class KAT {
     this.name = name;
 
     /**
-     * The extractor object for getting show data on torrents.
+     * The extractor object for getting anime data on torrents.
      * @type {Extractor}
      */
     this._extractor = new Extractor(this.name, new KatAPI({ debug }), debug);
@@ -36,7 +36,7 @@ export default class KAT {
   /**
    * Returns a list of all the inserted torrents.
    * @param {Object} provider - The provider to query https://kat.cr/.
-   * @returns {Show[]} - A list of scraped shows.
+   * @returns {Anime[]} - A list of scraped anime shows.
    */
   async search(provider) {
     try {
@@ -44,12 +44,11 @@ export default class KAT {
       provider.query.page = 1;
       provider.query.verified = 1;
       provider.query.adult_filter = 1;
-      provider.query.category = "tv";
-      provider.query.language = "en";
+      provider.query.category = "english-translated";
 
       return await this._extractor.search(provider);
     } catch (err) {
-      return this._util.onError(err);
+      this._util.onError(err);
     }
   }
