+++ src/providers/helpers/ShowHelper.js
@@ -2,7 +2,7 @@
 import asyncq from "async-q";
 
 import Show from "../../models/Show";
-import Util from "../../util";
+import Util from "../../Util";
 import { fanart, trakt, tmdb, tvdb } from "../../config/constants";
 
 /** Class for saving shows. */
