--- src/test.js
@@ -1,3 +1,0 @@
-import animes from '../tmp/animes.json';
-
-console.log(JSON.parse(animes));
